fix(ux): a11y correct description of People tab
- Used presentation role to deliver
intention of visible numeric value next to
the (en) label "People", though this role
have not removed number from A11Y
Tree, it still got pronounced, therefore
`aria-hidden` attribute is used.

- With number being only visually meaningful
we have to deliver missing message to visually
impared people, for this we created container
span that hold explanation of visual numeric value.

- Since in the conference there can be
at least one participant, we created two
separate translations that pluralize
message.

- We connect to the explanation via `aria-describedby`
attribute which causes NVDA to pronounce it after
element tab resulting in a string like
"People, tab, You are the only participant in the conference., 2 of 3"
instead of "People You are the only participant in the conference., tab, 2 of 3"