refactor: add microphone state caching
This changes nothing in the behaviour of the app.
In the future we can add a configuration option for the following.
When the user presses the whisper hotkey the microphone state can be cached.
When the user then releases the whisper hotkey the microphone state is set to the previous state.